Conflicts:
	.travis.yml
	src/Makefile.am
	src/suricata.c
	src/timemachine-heap.c
	src/timemachine-heap.h
	src/timemachine-packet.h
	src/timemachine.c
	src/timemachine.h
	src/tm-modules.c
	src/util-path.c
	src/util-path.h
	suricata.yaml.in

+++ src/util-file.c
@@ -873,7 +873,7 @@ void FileDisableStoringForTransaction(Flow *f, uint8_t direction, uint64_t tx_id
  *  \param fc file store
  *  \param file_id the file's id
  */
-void FileStoreFileById(FileContainer *fc, uint16_t file_id)
+void FileStoreFileById(FileContainer *fc, uint32_t file_id)
 {
     File *ptr = NULL;
 
@@ -888,7 +888,7 @@ void FileStoreFileById(FileContainer *fc, uint16_t file_id)
     }
 }
 
-void FileStoreAllFilesForTx(FileContainer *fc, uint16_t tx_id)
+void FileStoreAllFilesForTx(FileContainer *fc, uint64_t tx_id)
 {
     File *ptr = NULL;
